requirements.txt
streamlit
textblob
newspaper3k
nltk
PyPDF2
python-docx